Move command files to src/processes so they can be tested
